Create processAlldata.R
After tracking analysis (e.g., deeplabcut), we would obtain a lot of csv files. 
It is often needed to combine all data for your analysis. Also, we may need some processing to get variables which we will analyze, and the processing would be applied to each file.
This is a template to do these in R.